Ruler working
Cleaned some code
Divided in several .py files
